feat: Implement catalog navigation through links
The link is passed as a query parameter in the website URL, and the corresponding url is retrieved from the previously browsed page's links. This previously browsed page is stored in a session scope bean injected in the controller. In case of server crash or restart, the navigation session is lost and the controller redirects to the default catalog page.